WebApi/Application/GenreOperations/Commands/CreateGenre/CreateGenreCommand.cs
using System;
using System.Linq;
using AutoMapper;
using WebApi.DBOperations;
using WebApi.Entities;

namespace WebApi.Application.GenreOperations.Commands.CreateGenre
{
    public class CreateGenreCommand
    {
        public CreateGenreModel Model { get; set; }
        private readonly IMovieStoreDbContext _dbContext;
        private readonly IMapper _mapper;

        public CreateGenreCommand(IMovieStoreDbContext dbContext, IMapper mapper)
        {
            _dbContext = dbContext;
            _mapper = mapper;
        }

        public void Handle()
        {
            var genre=_dbContext.Genres.SingleOrDefault(
                        genre=> genre.GenreName.ToLower()==Model.GenreName.ToLower());
            if(genre is not null)
            throw new InvalidOperationException("Film türü zaten mevcut.");

            genre = _mapper.Map<Genre>(Model);
            _dbContext.Genres.Add(genre);
            _dbContext.SaveChanges();
        }
    }

    public class CreateGenreModel
    {
        public string GenreName { get; set; }
    }
}